requirements.txt
langchain-groq==0.1.9
langchain-core==0.2.38
python-dotenv==1.0.0